Add files via upload
n this project, I aimed to build and program a self-balancing robot within the 
Coppelia Simulation software. By using Coppelia Sim., a new door opens for 
designing, programming, and constructing any robot we like without any hurdles 
such as buying and waiting for components during the current scenario where 
the world is affected by the coronavirus pandemic. The construction and 
programming of the robot occurred in a virtual scene environment of Coppelia 
Simulation with the ease of working in any physical environment. At the end of 
the project, I was able to build a self-balancing robot that withstood external 
disturbances and balanced itself successfully.